Add files via upload
Cluster analysis is a multivariate tool for grouping or clustering items or observations based on how closely they are related. The primary goal of this project is to conduct cluster analysis with the chatterjee-price attitude dataset which can be found in datasets package in R. The data is aggregated from a survey of clerical employees which contain 35 employees for 30 randomly selected departments \citep{R_Datasets}. Non-hierarchical (k-means) method is used in this report to determine the distinct group in the Chatterjee-price attitude data.